refactor(frontend): shared downloadBlob/downloadJson util; ScenarioEditor uses templateApi
- utils/download.ts replaces the inline Blob/createObjectURL sequence
  (ProfilesTab's local copy moved there; ScenarioList's inline copy removed)
- ScenarioEditor fetches templates via templateApi.list()/get() instead of
  raw api.get with hand-rolled response generics (per the services-layer
  convention)
